+++ lib/ui/view/profile/profile_user_info_view.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:pocket_pose/config/app_color.dart';
 import 'package:pocket_pose/data/entity/response/profile_response.dart';
 import 'package:pocket_pose/ui/screen/profile/follow_tab_screen.dart';
-import 'package:pocket_pose/ui/frame/profile/profile_buttons_frame.dart';
+import 'package:pocket_pose/ui/view/profile/profile_buttons_view.dart';
 import 'package:pocket_pose/ui/widget/page_route_with_animation.dart';
 
 // ignore: must_be_immutable
