+++ lib/ui/view/video/video_right_view.dart
